+++ src/components/button/button.component.tsx
@@ -1,14 +1,15 @@
 import type { ButtonHTMLAttributes, CSSProperties, FC } from "react";
 import React from "react";
 import styles from "./button.module.scss";
+import classNames from "classnames";
 
-export enum ButtonColors {
+enum ButtonColors {
   primary,
   accent,
   danger,
 }
 
-interface IProps {
+const Button: FC<{
   onClick?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
   color?: ButtonColors;
   rounded?: boolean;
@@ -16,31 +17,18 @@ interface IProps {
   size?: string | number;
   type?: ButtonHTMLAttributes<HTMLButtonElement>["type"];
   styles?: CSSProperties;
-}
-
-export const Button: FC<IProps> = (props) => {
-  const classList: string[] = [props.className ?? "", styles.button];
-
-  switch (props.color) {
-    case ButtonColors.primary:
-      classList.push(styles.primary);
-      break;
-    case ButtonColors.accent:
-      classList.push(styles.accent);
-      break;
-    case ButtonColors.danger:
-      classList.push(styles.danger);
-      break;
-  }
-
-  if (props.rounded) {
-    classList.push(styles.rounded);
-  }
-
+  flat?: boolean;
+}> = (props) => {
   return (
     <button
       type={props.type ?? "button"}
-      className={classList.join(" ")}
+      className={classNames(props.className, styles.button, {
+        [styles.primary]: props.color === ButtonColors.primary,
+        [styles.accent]: props.color === ButtonColors.accent,
+        [styles.danger]: props.color === ButtonColors.danger,
+        [styles.flat]: props.flat,
+        [styles.rounded]: props.rounded,
+      })}
       onClick={props.onClick}
       style={{ ...props.styles, fontSize: props.size }}
     >
@@ -48,3 +36,5 @@ export const Button: FC<IProps> = (props) => {
     </button>
   );
 };
+
+export { Button, ButtonColors };
