+++ src/components/button/button.module.scss
@@ -1,19 +1,21 @@
-@use "../styles/variables";
-@use "../styles/typography";
+@use "../../styles/variables";
+@use "../../styles/typography";
 
 .button {
   @include typography.light;
-
+  background: none;
   border-radius: 100px;
   border: none;
   padding: 1rem;
+
+  cursor: pointer;
 
   &:focus {
     outline: none;
   }
 
   &:active {
-    box-shadow: inset 0px 0px 5px 1px rgba(0, 0, 0, 0.25);
+    box-shadow: inset 0 0 5px 1px rgba(0, 0, 0, 0.25);
   }
 }
 
@@ -37,4 +39,8 @@
   display: flex;
   justify-content: center;
   align-items: center;
+}
+
+.flat {
+
 }
