+++ src/components/button/index.ts
@@ -1,0 +1,1 @@
+export { Button, ButtonColors } from "./button.component";
