+++ src/components/index.ts
@@ -3,3 +3,4 @@ export { Modal } from "./modal";
 export { Overlay } from "./overlay";
 export { Profile } from "./profile";
 export { Avatar } from "./avatar";
+export { ButtonColors, Button } from "./button";
