+++ src/product/product-form.component.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useMemo, useState } from "react";
 import type { IProduct } from "./product.interface";
 import { ProductPriorityEnum } from "./product-priority.enum";
 import { InputText } from "../components/input-text.component";
-import { Button, ButtonColors } from "../components/button.component";
+import { Button, ButtonColors } from "../components";
 import type { SelectOption } from "../components/select.component";
 import { Select } from "../components/select.component";
 import { InputPrice } from "../components/input-price.component";
